Refactor cached discovery client
