+++ staging/src/k8s.io/client-go/discovery/cached/client_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package memcachediscovery
+package cached
 
 import (
 	"errors"
